Function to return IPAInputTestRow::random
